pyproject.toml
[tool.poetry]
name = "doodler"
version = "0.1.0"
description = "doodle game"
authors = ["sonduong305@gmail.com"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "0.60.1"
uvicorn = "^0.11.1"
pydantic = "^1.3"
requests = "^2.22.0"
loguru = "^0.4.0"
joblib = "^0.15.1"
aiofiles = "^0.5.0"
opencv-python = "^4.3.0"
Pillow = "^7.2.0"
olefile = "^0.46"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^1.4.4"
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
